pyproject.toml
[project]
name = "my-pydantic-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit",
    "pydantic-ai",
    "anthropic",
    "python-dotenv",
]
